Fixed Tab Fluidity
