+++ src/Main.java
@@ -1,10 +1,16 @@
 package src;
 
+import com.jfoenix.controls.JFXButton;
 import javafx.application.Application;
+import javafx.event.ActionEvent;
+import javafx.event.Event;
+import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
+
+import java.io.IOException;
 
 public class Main extends Application {
 
@@ -13,10 +19,10 @@ public class Main extends Application {
   @Override
   public void start(Stage primaryStage) throws Exception {
     setPrimaryStage(primaryStage);
-    Parent root = FXMLLoader.load(getClass().getResource("settings.fxml"));
+    Parent root = FXMLLoader.load(getClass().getResource("dashboard.fxml"));
     //primaryStage.setResizable(false);
     primaryStage.setTitle("Tutoring App");
-    primaryStage.setScene(new Scene(root, 600, 400));
+    primaryStage.setScene(new Scene(root, 600, 440));
     primaryStage.show();
 
   }
