+++ src/signin.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" ="false">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="src.signinController">
    <children>
       <Text layoutX="171.0" layoutY="69.0" strokeType="OUTSIDE" strokeWidth="0.0" text="Gutterboyz and Son Tutoring Center" textAlignment="CENTER" wrappingWidth="258.3359375">
          <font>
@@ -18,7 +18,7 @@
       </Text>
       <JFXTextField layoutX="172.0" layoutY="175.0" prefHeight="25.0" prefWidth="258.0" promptText="Username" />
       <JFXPasswordField layoutX="171.0" layoutY="228.0" prefHeight="25.0" prefWidth="258.0" promptText="Password" />
-      <JFXButton buttonType="RAISED" layoutX="378.0" layoutY="278.0" style="-fx-background-color: #ffffff;" text="Sign In" />
+      <JFXButton onAction="#button1Pressed" buttonType="RAISED" layoutX="378.0" layoutY="278.0" style="-fx-background-color: #ffffff;" text="Sign In" />
       <Hyperlink layoutX="200.0" layoutY="343.0" text="Would you like to make an account?" />
       <Label layoutX="173.0" layoutY="326.0" prefHeight="17.0" prefWidth="258.0" style="-fx-text-fill: #cc0000;" text="" />
    </children>
