+++ src/signinController.java
@@ -2,9 +2,14 @@ package src;
 
 import java.io.IOException;
 import javafx.event.ActionEvent;
+import javafx.event.Event;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
+import javafx.stage.Stage;
 
 
 public class signinController {// extends JSONObjectFactory{
@@ -15,8 +20,18 @@ public class signinController {// extends JSONObjectFactory{
 
     @FXML Label errortext;
 
-    //@FXML
-    /*private void button1Pressed(ActionEvent event) throws IOException {
+    @FXML
+
+    private void button1Pressed(ActionEvent event) throws IOException {
+            // System.out.println("in button1pressed");
+
+            Stage stage = Main.getPrimaryStage();
+
+            Parent root = FXMLLoader.load(getClass().getResource("dashboard.fxml"));
+
+            stage.setScene(new Scene(root, 600, 400));
+            stage.show();
+
         JSONObjectFactory JSONFile = new JSONObjectFactory();
         if (JSONFile.LoginValidation(usernametxt.getText(), passwordtxt.getText()) == Boolean.TRUE) {
             errortext.setText("Success");
@@ -29,5 +44,5 @@ public class signinController {// extends JSONObjectFactory{
 
 
     }
-    */
+
 }
